#12 fix bug in determining poets by century - Changing poet routes to the new architecture (Service-Repository pattern)
