+++ src/main/kotlin/Application.kt
@@ -6,10 +6,14 @@ import io.ktor.server.netty.*
 import io.ktor.server.routing.*
 import mobin.shabanifar.models.*
 import mobin.shabanifar.models.favorite.Favorite
+import mobin.shabanifar.models.poet.Poet
 import mobin.shabanifar.plugins.configureSerialization
 import mobin.shabanifar.repository.FavoriteRepository
+import mobin.shabanifar.repository.PoetRepository
 import mobin.shabanifar.routes.favoriteRoutes
+import mobin.shabanifar.routes.poetRoutes
 import mobin.shabanifar.service.FavoriteService
+import mobin.shabanifar.service.PoetService
 import org.jetbrains.exposed.sql.Database
 import org.jetbrains.exposed.sql.SchemaUtils
 import org.jetbrains.exposed.sql.transactions.transaction
@@ -23,10 +27,13 @@ fun Application.module() {
     configureSerialization()
     val favoriteRepository = FavoriteRepository()
     val favoriteService = FavoriteService(favoriteRepository)
+    val poetRepository = PoetRepository()
+    val poetService = PoetService(poetRepository)
 
     routing {
         createRoute()
         favoriteRoutes(favoriteService)
+        poetRoutes(poetService)
     }
 }
