+++ src/main/kotlin/Routing.kt
@@ -12,30 +12,6 @@ import org.jetbrains.exposed.sql.transactions.transaction
 fun Route.createRoute() {
     route("/api") {
 
-        get("/getPoetsByCentury") {
-            try {
-                // Extract query parameter
-                val century = call.request.queryParameters["century"]?.toIntOrNull()
-
-                // Validate query parameter
-                if (century == null) {
-                    call.respond(
-                        HttpStatusCode.BadRequest,
-                        "Missing or invalid query parameter: 'century' must be a valid integer"
-                    )
-                    return@get
-                }
-
-                // Fetch poets for the specified century
-                val poetsByCentury = getPoetsByCentury(century)
-
-                // Respond with the result
-                call.respond(HttpStatusCode.OK, poetsByCentury)
-            } catch (e: Exception) {
-                call.respond(HttpStatusCode.InternalServerError, "An error occurred: ${e.message}")
-            }
-        }
-
         get("/getPoemsOfCategory") {
             try {
                 val poetName = call.request.queryParameters["poetName"]
@@ -67,40 +43,6 @@ fun Route.createRoute() {
 
                 // Respond with the result
                 call.respond(HttpStatusCode.OK, randomVerse)
-            } catch (e: Exception) {
-                call.respond(HttpStatusCode.InternalServerError, "An error occurred: ${e.message}")
-            }
-        }
-
-        get("/getWorksOfPoet") {
-            try {
-                // Extract query parameter
-                val poetName = call.request.queryParameters["poetName"]
-
-                // Validate query parameter
-                if (poetName == null) {
-                    call.respond(
-                        HttpStatusCode.BadRequest,
-                        "Missing or invalid query parameter: 'poetName' must be a valid string"
-                    )
-                    return@get
-                }
-
-                val getWorksOfPoet = getWorksOfPoet(poetName)
-
-                // Respond with the result
-                call.respond(HttpStatusCode.OK, getWorksOfPoet)
-            } catch (e: Exception) {
-                call.respond(HttpStatusCode.InternalServerError, "An error occurred: ${e.message}")
-            }
-        }
-
-        get("/getTop8FamousPoets") {
-            try {
-                val getTop8FamousPoets = getTop8FamousPoets()
-
-                // Respond with the result
-                call.respond(HttpStatusCode.OK, getTop8FamousPoets)
             } catch (e: Exception) {
                 call.respond(HttpStatusCode.InternalServerError, "An error occurred: ${e.message}")
             }
