+++ src/main/kotlin/models/PoetImage.kt
@@ -2,6 +2,7 @@ package mobin.shabanifar.models
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
+import mobin.shabanifar.models.poet.Poet
 import org.jetbrains.exposed.sql.Table
 import org.jetbrains.exposed.sql.select
 import org.jetbrains.exposed.sql.transactions.transaction
