+++ src/main/kotlin/models/poet/Category.kt
@@ -1,0 +1,6 @@
+package mobin.shabanifar.models.poet
+
+data class Category(
+    val text: String?,
+    val url: String?
+)
