+++ src/main/kotlin/models/poet/FamousPoet.kt
@@ -1,0 +1,7 @@
+package mobin.shabanifar.models.poet
+
+data class FamousPoet(
+    val id: Int,
+    val name: String?,
+    val description: String?
+)
