+++ src/main/kotlin/models/poet/PoetWithBirthYear.kt
@@ -1,0 +1,9 @@
+package mobin.shabanifar.models.poet
+
+data class PoetWithBirthYear(
+    val id: Int,
+    val name: String,
+    val catId: Int?,
+    val description: String?,
+    val birthYear: Int?
+)
